Move `ColumnPositions` utility
